src/test/java/org/gd/leetcode/p0662/SolutionTest.java
package org.gd.leetcode.p0662;

import org.gd.leetcode.common.TreeNode;
import org.junit.jupiter.api.*;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

/**
 * Test for {@link Solution}
 *
 * @author Horkhover Dmytro
 * @since 2020-07-18
 */
class SolutionTest {

    private static Stream<Arguments> args() {
        return Stream.of(

                Arguments.of(TreeNode.of(
                        0, 0, 0, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0,
                        null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null,
                        null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null,
                        0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0,
                        null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null,
                        null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null,
                        0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0,
                        null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null,
                        null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null,
                        0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0,
                        null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null,
                        null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null, null,
                        0, 0, null, null, 0, 0, null, null, 0, 0, null, null, 0, 0, null), 2),

                Arguments.of(TreeNode.of(1, 1, 1, 1, null, null, 1, 1, null, null, 1), 8),
                Arguments.of(TreeNode.of(1, 3, 2, 5), 2),
                Arguments.of(TreeNode.of(1, 3, 2, 5, 3, null, 9), 4)
        );
    }

    @ParameterizedTest
    @MethodSource("args")
    @DisplayName("WidthOfBinaryTree")
    void test_WidthOfBinaryTree(TreeNode root, int expected) {
        assertEquals(expected, new Solution().widthOfBinaryTree(root));
    }
}